pyproject.toml
[tool.poetry]
name = "cliver"
version = "0.1.0"
description = "Cliver, A FastAPI with Django ORM Template"
authors = ["Uhtred M"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
argon2-cffi = "^23.1.0"
django = "^5.0.3"
asgi-correlation-id = "^4.3.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.44.0"}
python-jose = "^3.3.0"
rich = "^13.7.1"
typer = "^0.12.0"
email-validator = "^2.1.1"


[tool.poetry.group.dev.dependencies]
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
types-python-jose = "^3.3.4.20240106"
psycopg2-binary = "^2.9.9"
pytest = "^8.1.1"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "cliver.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
